moves all nested files in dashboard to root and all tests to one folder
