+++ src/reducers.js
@@ -1,8 +1,8 @@
 import { combineReducers } from 'redux';
 import user from './components/authentication/user.reducer';
 import authentication from './components/authentication/authentication.reducer';
-import decadevs from './components/dashboard/decadevs-reducers/decadevs.reducer';
-import interviewDetails from './components/dashboard/scheduleInterview/scheduleInterview.reducer';
+import decadevs from './components/dashboard/decadevs.reducer';
+import interviewDetails from './components/dashboard/scheduleInterview.reducer';
 import hirer from './components/InviteHirer/inviteHirer.reducer';
 
 export default combineReducers({
